pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-quality-monitor"
version = "0.1.0"
description = "RAG quality monitoring and assurance platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Onurcan Donmezer" },
]
keywords = ["rag", "quality", "monitoring", "llm", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "streamlit>=1.30",
    "pandas>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/onurcandnmz/rag-quality-monitor"
Repository = "https://github.com/onurcandnmz/rag-quality-monitor"
Issues = "https://github.com/onurcandnmz/rag-quality-monitor/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
omit = ["src/app.py"]
